backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
moviepy==1.0.3
Pillow==10.3.0
requests==2.32.2
python-dotenv==1.0.1
edge-tts==6.1.12
aiofiles==23.2.1
aiohttp==3.9.5
